pyproject.toml
[project]
name = "ai-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=0.3.0",
    "docker>=6.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=6.9.2",
    "pymupdf>=1.27.2.2",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "accelerate>=1.13.0",
    "zeroentropy>=0.1.0a11",
]
